+++ saas-b2b-b2c/frontend/jest.config.js
@@ -24,10 +24,11 @@ module.exports = {
   ],
   coverageThreshold: {
     global: {
-      branches: 61,
+      branches: 62,
       functions: 70,
       lines: 77,
-      statements: 77,
+      statements: 78,
     },
+  },
   },
 };